moved Token, TokenData and TokenTypes to _tokens.py module
